Add mapper for DTO conversion
